ikhal: shift start/end dt on invalid input
When the user inserting a start date > the end date (or the otherwise
around) we shift the other datetime so that the original length of the
event stays the same.

You can still escape this by using the key-combination (default meta +
enter) to save, which may save in invalid VEVENT. This needs to be
fixed.

This fixes some parts of #335.